# packages/contracts/openapi/career-agent.openapi.yaml
openapi: 3.0.3
info:
  title: Career Agent API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/profile:
    get:
      summary: List Student Profiles
      responses:
        "200":
          description: Student profile list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListStudentProfilesResponse"
    post:
      summary: Create Student Profile From Form Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentProfileRequest"
      responses:
        "201":
          description: Created profile with completeness and competitiveness scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/profile/resume:
    post:
      summary: Create Student Profile From Resume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, target_role]
              properties:
                file:
                  type: string
                  format: binary
                  description: 简历文件
                target_role:
                  type: string
                  minLength: 1
                name:
                  type: string
                parse_mode:
                  type: string
                  enum: [strict, tolerant]
                  default: tolerant
      responses:
        "201":
          description: Created student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Resume parse failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/career-paths/jobs/{job_id}:
    get:
      summary: Get Career Path Graph For Job
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
            minimum: 1
        - in: query
          name: student_profile_id
          schema:
            type: integer
            minimum: 1
        - in: query
          name: depth
          schema:
            type: integer
            minimum: 1
            maximum: 3
            default: 2
      responses:
        "200":
          description: Career path graph result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPathGraphResponse"
        "404":
          description: Job or canonical career role not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Neo4j graph service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v2/jobs/pipeline/run:
    post:
      summary: Run Jobs Intelligence Pipeline
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [cleanse_agent_portraits]
                  default: cleanse_agent_portraits
      responses:
        "202":
          description: Pipeline task accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPipelineTaskRecord"
  /api/v2/jobs/pipeline/tasks/{task_id}:
    get:
      summary: Get Jobs Pipeline Task
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Pipeline task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPipelineTaskRecord"
  /api/v2/jobs/pipeline/tasks/{task_id}/failures:
    get:
      summary: List Pipeline Task Failures
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
            minimum: 1
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Pipeline task failure list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPipelineFailureListResponse"
  /api/v2/jobs/pipeline/retry-queue:
    get:
      summary: List Pipeline Retry Queue
      parameters:
        - in: query
          name: task_id
          schema:
            type: integer
            minimum: 1
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, done, failed]
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Pipeline retry queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPipelineRetryQueueListResponse"
  /api/v2/career-paths/jobs/{job_id}:
    get:
      summary: Get Career Path Graph V2
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
            minimum: 1
        - in: query
          name: depth
          schema:
            type: integer
            minimum: 1
            maximum: 3
            default: 2
      responses:
        "200":
          description: Career path graph result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPathGraphResponse"
...
components:
  schemas:
    DimensionScores:
      type: object
      required:
        [base_requirements, professional_skills, professional_quality, development_potential]
      properties:
        base_requirements:
          type: number
          minimum: 0
          maximum: 100
        professional_skills:
          type: number
          minimum: 0
          maximum: 100
        professional_quality:
          type: number
          minimum: 0
          maximum: 100
        development_potential:
          type: number
          minimum: 0
          maximum: 100
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          nullable: true
        trace_id:
          type: string
...
    JobPipelineTaskRecord:
      type: object
      required:
        [
          id,
          mode,
          status,
          total_jobs,
          processed_jobs,
          success_profiles,
          failed_profiles,
          graph_nodes,
          graph_edges,
          family_count,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: integer
        mode:
          type: string
          enum: [cleanse_agent_portraits]
        status:
          type: string
          enum: [queued, running, success, degraded, failed]
        total_jobs:
          type: integer
        processed_jobs:
          type: integer
        success_profiles:
          type: integer
        failed_profiles:
          type: integer
        graph_nodes:
          type: integer
        graph_edges:
          type: integer
        family_count:
          type: integer
        message:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          nullable: true
          format: date-time
        finished_at:
          type: string
          nullable: true
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobPipelineFailureRecord:
      type: object
      required:
        [id, task_id, job_id, stage, error_code, error_message, attempts, retryable, created_at]
      properties:
        id:
          type: integer
        task_id:
          type: integer
        job_id:
          type: integer
        stage:
          type: string
        error_code:
          type: string
        error_message:
          type: string
        attempts:
          type: integer
        retryable:
          type: boolean
        created_at:
          type: string
          format: date-time
    JobPipelineFailureListResponse:
      type: object
      required: [total, items]
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/JobPipelineFailureRecord"
    JobPipelineRetryQueueRecord:
      type: object
      required: [id, task_id, job_id, stage, status, attempts, next_run_at, created_at, updated_at]
      properties:
        id:
          type: integer
        task_id:
          type: integer
        job_id:
          type: integer
        stage:
          type: string
        status:
          type: string
          enum: [pending, processing, done, failed]
        attempts:
          type: integer
        next_run_at:
          type: string
          format: date-time
        last_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobPipelineRetryQueueSummary:
      type: object
      required: [pending, processing, done, failed, latest_errors]
      properties:
        pending:
          type: integer
        processing:
          type: integer
        done:
          type: integer
        failed:
          type: integer
        latest_errors:
          type: array
          items:
            type: string
    JobPipelineRetryQueueListResponse:
      type: object
      required: [total, items, summary]
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/JobPipelineRetryQueueRecord"
        summary:
          $ref: "#/components/schemas/JobPipelineRetryQueueSummary"
    CreateStudentProfileRequest:
      type: object
      required: [name, target_role, skills]
      properties:
        name:
          type: string
        target_role:
          type: string
        education_level:
          type: string
        major:
          type: string
        graduation_year:
          type: integer
          minimum: 2000
          maximum: 2100
        skills:
          type: array
          minItems: 1
          items:
            type: string
        certificates:
          type: array
          items:
            type: string
        experience:
          type: object
          properties:
            internship_count:
              type: integer
              minimum: 0
            project_count:
              type: integer
              minimum: 0
            competition_count:
              type: integer
              minimum: 0
        self_assessment:
          type: object
          properties:
            communication:
              type: integer
              minimum: 1
              maximum: 5
            learning:
              type: integer
              minimum: 1
              maximum: 5
            stress_tolerance:
              type: integer
              minimum: 1
              maximum: 5
            innovation:
              type: integer
              minimum: 1
              maximum: 5
        personal_summary:
          type: string
    StudentProfile:
      type: object
      required:
        [
          id,
          source_type,
          source_digest,
          name,
          target_role,
          skills,
          dimension_scores,
          completeness_score,
          competitiveness_score,
          created_at,
        ]
      properties:
        id:
          type: integer
        source_type:
          type: string
          enum: [manual, resume]
        source_digest:
          type: string
        name:
          type: string
        target_role:
          type: string
        education_level:
          type: string
          nullable: true
        major:
          type: string
          nullable: true
        graduation_year:
          type: integer
          nullable: true
        skills:
          type: array
          items:
            type: string
        certificates:
          type: array
          items:
            type: string
        experience:
          type: object
          properties:
            internship_count:
              type: integer
            project_count:
              type: integer
            competition_count:
              type: integer
        self_assessment:
          type: object
          properties:
            communication:
              type: integer
            learning:
              type: integer
            stress_tolerance:
              type: integer
            innovation:
              type: integer
        dimension_scores:
          $ref: "#/components/schemas/DimensionScores"
        completeness_score:
          type: number
        competitiveness_score:
          type: number
        missing_items:
          type: array
          items:
            type: string
        personal_summary:
          type: string
          nullable: true
        summary:
          type: string
        created_at:
          type: string
          format: date-time
    ListStudentProfilesResponse:
      type: object
      required: [total, items]
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/StudentProfile"
...
    CareerPathNode:
      type: object
      required:
        [
          id,
          job_id,
          role_key,
          title,
          description,
          family,
          level,
          aliases,
          typical_skills,
          category,
          is_target,
        ]
      properties:
        id:
          type: string
        job_id:
          type: integer
          nullable: true
        role_key:
          type: string
        title:
          type: string
        description:
          type: string
        family:
          type: string
        level:
          type: integer
        aliases:
          type: array
          items:
            type: string
        typical_skills:
          type: array
          items:
            type: string
        category:
          type: string
          enum: [target, promotion, transition]
        is_target:
          type: boolean
    CareerPathEdge:
      type: object
      required:
        [
          id,
          source,
          target,
          relation_type,
          reason,
          required_skills,
          gap_skills,
          transition_cost,
          direction_label,
          score,
        ]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        relation_type:
          type: string
          enum: [promotion, transition]
        reason:
          type: string
        required_skills:
          type: array
          items:
            type: string
        gap_skills:
          type: array
          items:
            type: string
        transition_cost:
          type: string
          enum: [low, medium, high]
        direction_label:
          type: string
        score:
          type: integer
    CareerRouteStep:
      type: object
      required:
        [
          node_id,
          role_key,
          title,
          relation_type,
          reason,
          required_skills,
          transition_cost,
          gap_skills,
        ]
      properties:
        node_id:
          type: string
        role_key:
          type: string
        title:
          type: string
        relation_type:
          type: string
          nullable: true
          enum: [promotion, transition]
        reason:
          type: string
          nullable: true
        required_skills:
          type: array
          items:
            type: string
        transition_cost:
          type: string
          nullable: true
          enum: [low, medium, high]
        gap_skills:
          type: array
          items:
            type: string
    CareerRouteRecommendation:
      type: object
      required: [route_id, route_type, title, summary, suitability_score, missing_skills, steps]
      properties:
        route_id:
          type: string
        route_type:
          type: string
          enum: [promotion, transition]
        title:
          type: string
        summary:
          type: string
        suitability_score:
          type: integer
        missing_skills:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/CareerRouteStep"
    CareerPathGraphResponse:
      type: object
      required:
        [
          job_id,
          job_title,
          depth,
          target_node_id,
          nodes,
          edges,
          promotion_routes,
          transition_routes,
        ]
      properties:
        job_id:
          type: integer
        job_title:
          type: string
        depth:
          type: integer
        target_node_id:
          type: string
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CareerPathNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/CareerPathEdge"
        promotion_routes:
          type: array
          items:
            $ref: "#/components/schemas/CareerRouteRecommendation"
        transition_routes:
          type: array
          items:
            $ref: "#/components/schemas/CareerRouteRecommendation"